requirements.txt
Django==5.0.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.26.5
Pillow>=10.1.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.1
django-filter>=23.5
pytest>=7.4.3
pytest-django>=4.7.0
djoser>=2.2.2
pytz>=2023.3